Refactored DictyCell

+++ Dicty/Dicty/LanguageModel.swift
@@ -4,10 +4,10 @@ import CoreData
 
 class LanguageModel {
     let name: String
-    let id: Int16
+    let id: Int
     let shortName: TranslatorSupportedLanguage
 
-    init(_ name: String, _ id: Int16, _ shortName: TranslatorSupportedLanguage) {
+    init(_ name: String, _ id: Int, _ shortName: TranslatorSupportedLanguage) {
         self.name = name
         self.id = id
         self.shortName = shortName
@@ -15,7 +15,7 @@ class LanguageModel {
 
     convenience init?(obj: NSManagedObject) {
         guard let name = obj.value(forKey: T.Languages.Name) as? String,
-              let id = obj.value(forKey: T.Languages.Id) as? Int16,
+              let id = obj.value(forKey: T.Languages.Id) as? Int,
               let shortNameStr = obj.value(forKey: T.Languages.ShortName) as? String,
               let shortName = TranslatorSupportedLanguage(rawValue: shortNameStr)
         else {
